extract methods to abstract class
