+++ src/Splitter/TestsSplitterTask.php
@@ -4,12 +4,10 @@ namespace Codeception\Task\Splitter;
 use Codeception\Lib\Di;
 use Codeception\Test\Cest;
 use Codeception\Test\Descriptor as TestDescriptor;
-use Codeception\Test\Loader as TestLoader;
 use Exception;
 use PHPUnit\Framework\DataProviderTestSuite;
 use PHPUnit\Framework\TestCase;
 use ReflectionObject;
-use Robo\Exception\TaskException;
 
 /**
  * Loads all tests into groups and saves them to groupfile according to pattern.
@@ -27,12 +25,7 @@ class TestsSplitterTask extends TestsSplitter
 {
     public function run()
     {
-        if (!$this->doCodeceptLoaderExists()) {
-            throw new TaskException(
-                $this,
-                'This task requires Codeception to be loaded. Please require autoload.php of Codeception'
-            );
-        }
+        $this->claimCodeceptionLoaded();
         $tests = $this->loadTests();
 
         $this->printTaskInfo('Processing ' . count($tests) . ' tests');
@@ -148,32 +141,5 @@ class TestsSplitterTask extends TestsSplitter
         }
 
         return true;
-    }
-
-    /**
-     * @return bool
-     */
-    protected function doCodeceptLoaderExists(): bool
-    {
-        return class_exists(TestLoader::class);
-    }
-
-    /**
-     * @return TestLoader
-     */
-    protected function getTestLoader(): TestLoader
-    {
-        return new TestLoader(['path' => $this->testsFrom]);
-    }
-
-    /**
-     * @return array
-     */
-    protected function loadTests(): array
-    {
-        $testLoader = $this->getTestLoader();
-        $testLoader->loadTests($this->testsFrom);
-
-        return $testLoader->getTests();
     }
 }
